refactor: 파일구조 및 파일명 수정 / attachEventListener 호출 위치 변경
