+++ project_root/frontend/js/core/event.js
@@ -1,8 +1,7 @@
 import { navigateTo } from './router.js';
 import { gameSettings, renderGameOptions } from '../pages/settingOptions.js';
-import { renderSubgameMenu } from '../pages/subgame/subgameDisplay.js';
-import { startMatch, showResult } from '../pages/subgame/subgameMatch.js';
-import { attachEventListener } from '../pages/subgame/subgameButtonListener.js';
+import { renderSubgameMenu } from '../subgame/renderSubgameMenu.js';
+import { startMatch, showResult } from '../subgame/match.js';
 import { postEnrollment } from '../api/enroll.js';
 
 // click 이벤트 발생 시 유형에 따라 처리
@@ -26,9 +25,7 @@ export async function eventHandler(event) {
 
     if (target.matches('[data-action="renderSubgameMenu"]')) {
         const mode = target.getAttribute('data-mode');
-        renderSubgameMenu(mode).then(() => {
-            attachEventListener();
-        });
+        renderSubgameMenu(mode);
     }
 
     if (target.matches('[data-action="subgameSave"]')) {
