+++ project_root/frontend/js/subgame/didWin.js
